table is rendering but is repeating the first four indexes in the api return multiple times. Data is being displayed incorrectly
